refactor: reorganize components and extract barrel exports
- Reorganize components into `lists/` and `details/` subdirectories
- Rename MyCollection to CollectionSubjectList
- Rename ProgressViewer to ProgressGrid
- Move getCollectionTag to src/shared/utils.ts
- Restore translatedText to useAITranslate hook return values
- Create barrel exports (index.ts) for components/lists, components/details, and api
- Refactor project-wide imports to use path aliases (@/api, @/components/lists, etc.)
