+++ extensions/bangumi/src/components/details/index.ts
@@ -1,0 +1,3 @@
+export { default as CharacterDetail } from "./CharacterDetail"
+export { default as EpisodeDetail } from "./EpisodeDetail"
+export { default as SubjectDetail } from "./SubjectDetail"
