+++ extensions/bangumi/src/components/lists/index.ts
@@ -1,0 +1,4 @@
+export { default as CollectionSubjectList } from "./CollectionSubjectList"
+export { default as ProgressGrid } from "./ProgressGrid"
+export { default as RelationsList } from "./RelationsList"
+export { default as SubjectCharactersList } from "./SubjectCharactersList"
