+++ extensions/bangumi/src/tools/get-subject-details.ts
@@ -1,8 +1,7 @@
-import { bangumi } from "@/api/bangumi"
 import { withAccessToken } from "@raycast/utils"
-import { bangumiAuth } from "@/api/oauth"
 import { formatSubjectToMarkdown, formatInfoboxToMarkdown } from "./utils"
 import { getCollectionTag } from "@/shared/utils"
+import { bangumi, bangumiAuth } from "@/api"
 
 type Input = {
   /**
