+++ extensions/bangumi/src/tools/search-subjects.ts
@@ -1,7 +1,6 @@
-import { bangumi } from "@/api/bangumi"
 import { withAccessToken } from "@raycast/utils"
-import { bangumiAuth } from "@/api/oauth"
 import { formatSubjectToMarkdown } from "./utils"
+import { bangumi, bangumiAuth } from "@/api"
 
 type Input = {
   /**
